Moved all the tests back after reading an article
