+++ frontend/src/Dashboard/actions.async.unit.spec.js
@@ -1,6 +1,6 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import * as actions from '../../Dashboard/actions';
+import * as actions from './actions';
 import fetchMock from 'fetch-mock';
 import {expect} from 'chai';
